Gave every marshalled class a separate file in a new types folder
