+++ config/types/card_type_combination_generator.py
@@ -1,29 +1,6 @@
 from util.generators.constrained_product import ProductEQ, ProductLE
 
 
-class ReadOnlyAttrDict:
-    def __init__(self, **kwargs):
-        self.__dict__ = kwargs
-
-    def __iter__(self):
-        return iter(self.__dict__)
-
-    def items(self):
-        return self.__dict__.items()
-
-    def keys(self):
-        return self.__dict__.keys()
-
-    def values(self):
-        return self.__dict__.values()
-
-    def __getitem__(self, item):
-        return self.__dict__.__getitem__(item)
-
-    def __repr__(self):
-        return self.__dict__.__repr__()
-
-
 class CardTypeCombinationGenerator:
     __constraints = {
         'sum_eq': ProductEQ,
